requirements.txt.txt
streamlit
google-genai
elevenlabs